backend/requirements.txt
Flask
flask-cors
python-dotenv
requests
beautifulsoup4
google-generativeai
gunicorn
